requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
plotly==5.18.0
pandas==2.1.4
APScheduler==3.10.4
Flask-Mail==0.9.1